# openapi.yaml
openapi: 3.0.3
info:
  title: Suminregio API
  description: |
    API para dashboard ejecutivo de Suminregio Parker (ventas, CXC, inventario,
    contabilidad, consumos, análisis avanzado). Consume DuckDB snapshots.
  version: '9.1'
  contact:
    name: Guillermo Ragnar
servers:
  - url: https://suminregioparker-todoslosnegocios.onrender.com
    description: Producción (Render)
  - url: http://localhost:7000
    description: Desarrollo local
...
paths:
...
  /api/analytics/pareto:
    get:
      tags: [Analytics]
      summary: Análisis 80/20 de clientes o artículos
      parameters:
        - in: query
          name: dim
          schema: { type: string, enum: [cliente, articulo] }
        - in: query
          name: dias
          schema: { type: integer, default: 365 }
...
  /api/compare/temporal:
    get:
      tags: [Analytics]
      summary: Comparación MoM y YoY para una o más métricas
      parameters:
        - in: query
          name: metrics
          schema: { type: string, example: 'ventas_mes,cxc_total' }
  /api/search/global:
    get:
      tags: [Analytics]
      summary: Búsqueda unificada (clientes, vendedores, artículos, folios)
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
...
  /api/ai/chat-v2:
    post:
      tags: [AI]
      summary: Chat conversacional con memoria por sesión (Anthropic Claude)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message: { type: string }
                sessionId: { type: string }
                db: { type: string }
...